refactor: apply simplify-pass findings for email notifications
Use useSyncFromQuery for the email form draft instead of the
useEffect+setState anti-pattern the hook was created to replace, and
extract EmailSettingsForm plus the shared event catalog into their own
files. Replace the hand-rolled HTML escaper with html.escape and the
logo cache globals with lru_cache. Import email_service at module level
in webhook_service and call send_event_email bare since it guarantees
never-raises; the isolation test now fails at the SMTP layer to
exercise that guarantee for real. Drop the pointless port derivation in
the PUT handler, the duplicated metadata pre-check in
validate_base_url, and share the SyncThread test stub.
