+++ frontend/src/pages/settings/notificationEvents.ts
@@ -1,0 +1,8 @@
+// Notification event catalog shared by the email form and webhook editor.
+export const EVENT_OPTIONS: { value: string; label: string }[] = [
+  { value: 'Episode Processed', label: 'Episode Processed' },
+  { value: 'Episode Failed', label: 'Episode Failed' },
+  { value: 'Auth Failure', label: 'Auth Failure' },
+  { value: 'Limit Exceeded', label: 'Limit Exceeded' },
+  { value: 'Rate Limit Structural', label: 'Rate Limit Structural' },
+];
